Get basic happy path tests passing
Extracts the shortcut handler into its own callback for more descriptive
and testable logic.

The app and registerShortcutListener logic are now both coupled and
separate, so I may come back around to clean that up in the future, but
this allows for better unit testing so it may be okay to leave as-is.

To do:

* Sad path tests
* Test cleanup